+++ server/requirements/RequirementsChecker.php
@@ -656,7 +656,10 @@ class RequirementsChecker
             // Get the base path without the script name.
             $subBasePath = \craft\helpers\FileHelper::normalizePath(mb_substr(Craft::$app->getRequest()->getScriptFile(), 0, -mb_strlen(Craft::$app->getRequest()->getScriptUrl())));
 
-            if (mb_strpos($pathToTest, $subBasePath) !== false) {
+            // If $subBasePath === '', then both the craft folder and index.php are living at the root of the filesystem.
+            // Note that some web servers (Idea Web Server) can be configured with virtual roots so that PHP's realpath
+            // returns that instead of the actual root.
+            if ($subBasePath === '' || mb_strpos($pathToTest, $subBasePath) !== false) {
                 return true;
             }
         }